+++ modules/core/include/nt2/core/functions/common/blkdiag.hpp
@@ -45,7 +45,7 @@ namespace nt2 { namespace ext
       _2D ex0 = boost::proto::child_c<0>(a0).extent();
       sub_t pos = ind2sub(a0.extent(),p);
       i_t p0 = nt2::enumerate<i_t>(pos[0]);
-      if(pos[1] <= ex0[1]) // this is in the a0 stack
+      if(size_t(pos[1]) <= ex0[1]) // this is in the a0 stack
         {
           State pp = sub2ind(ex0, pos);
           return nt2::if_else( nt2::le(p0, nt2::splat<i_t>(ex0[0])),
