+++ modules/core/include/nt2/core/functions/scalar/putalong.hpp
@@ -29,7 +29,7 @@ namespace nt2 { namespace ext
     BOOST_FORCEINLINE result_type operator()(A0 const& a0, A1 const&) const
     {
       ext_t ext; 
-      for(int i=0; i < ext.size(); ++i) ext[i] = 1;
+      for(size_t i=0; i < ext.size(); ++i) ext[i] = 1;
       tab_t t(ext);
       t(1) = a0; 
       return t;
