+++ modules/core/unit/functions/center.cpp
@@ -52,8 +52,8 @@ NT2_TEST_CASE_TPL( center, (float)(double))//NT2_TYPES )
   nt2::table<T> y( nt2::of_size(8, 4) );
   nt2::table<T> sy, m, z;
 
-  for(int j=1;j<=size(y, 2);j++)
-    for(int i=1;i<=size(y, 1);i++)
+  for(size_t j=1;j<=size(y, 2);j++)
+    for(size_t i=1;i<=size(y, 1);i++)
       y(i,j) = i - j;
 
   sy = nt2::center(y, 2);
