+++ modules/core/unit/functions/dot.cpp
@@ -39,26 +39,26 @@ NT2_TEST_CASE_TPL( dot, (float)(double))//NT2_TYPES )
   nt2::table<T> sz;
 
 
-  for(int j=1;j<=size(sy, 2);j++)
-    for(int i=1;i<=size(sy, 1);i++)
+  for(size_t j=1;j<=size(sy, 2);j++)
+    for(size_t i=1;i<=size(sy, 1);i++)
       y(i,j) = i - j;
 
   sy = nt2::dot(y, y);
   sz = nt2::sum(nt2::multiplies(y, y));
-  for(int j=1;j<=size(y, 2);j++)
+  for(size_t j=1;j<=size(y, 2);j++)
       NT2_TEST_EQUAL(sz(j), sy(j));
   sy = nt2::dot(y, y, 1);
   sz = nt2::sum(nt2::multiplies(y, y), 1);
-  for(int j=1;j<=size(y, 2);j++)
+  for(size_t j=1;j<=size(y, 2);j++)
       NT2_TEST_EQUAL(sz(j), sy(j));
   sy = nt2::dot(y, y, 2);
   sz = nt2::sum(nt2::multiplies(y, y), 2);
-    for(int i=1;i<=size(y, 1);i++)
+    for(size_t i=1;i<=size(y, 1);i++)
       NT2_TEST_EQUAL(sz(i), sy(i));
   sy = nt2::dot(y, y, 3);
   sz = nt2::sum(nt2::multiplies(y, y), 3);
-  for(int j=1;j<=size(y, 2);j++)
-    for(int i=1;i<=size(y, 1);i++)
+  for(size_t j=1;j<=size(y, 2);j++)
+    for(size_t i=1;i<=size(y, 1);i++)
       NT2_TEST_EQUAL(sz(i, j), sy(i, j));
