+++ modules/core/unit/functions/mean.cpp
@@ -45,32 +45,32 @@ NT2_TEST_CASE_TPL( mean, (float)(double))//NT2_TYPES )
   sy2 = nt2::mean(y);
   display("sy", sy);
   display("sy2", sy2);
-  for(int j=1;j<=size(sy, 2);j++)
-    for(int i=1;i<=size(sy, 1);i++)
+  for(size_t j=1;j<=size(sy, 2);j++)
+    for(size_t i=1;i<=size(sy, 1);i++)
       NT2_TEST_EQUAL(sy(i,j), sy2(i, j));
 
   sy = nt2::sum(y, 1)/T(nt2::size(y, 1));
   sy2 = nt2::mean(y, 1);
   display("sy", sy);
   display("sy2", sy2);
-  for(int j=1;j<=size(sy, 2);j++)
-    for(int i=1;i<=size(sy, 1);i++)
+  for(size_t j=1;j<=size(sy, 2);j++)
+    for(size_t i=1;i<=size(sy, 1);i++)
       NT2_TEST_EQUAL(sy(i,j), sy2(i, j));
 
   sy = nt2::sum(y, 2)/T(nt2::size(y, 2));
   sy2 = nt2::mean(y, 2);
   display("sy", sy);
   display("sy2", sy2);
-  for(int j=1;j<=size(sy, 2);j++)
-    for(int i=1;i<=size(sy, 1);i++)
+  for(size_t j=1;j<=size(sy, 2);j++)
+    for(size_t i=1;i<=size(sy, 1);i++)
       NT2_TEST_EQUAL(sy(i,j), sy2(i, j));
 
   sy = nt2::sum(y, 3)/T(nt2::size(y, 3));
   sy2 = nt2::mean(y, 3);
   display("sy", sy);
   display("sy2", sy2);
-  for(int j=1;j<=size(sy, 2);j++)
-    for(int i=1;i<=size(sy, 1);i++)
+  for(size_t j=1;j<=size(sy, 2);j++)
+    for(size_t i=1;i<=size(sy, 1);i++)
       NT2_TEST_EQUAL(sy(i,j), sy2(i, j));
